Project files that passed all 7 tests
I used the motor controlled I made as I understood that more. I also used my ALU, as it was not passing with the one we had in the project prior. Also messed with SPI_Master (using option 2).  Have not checked the math with perl yet! However, as I mentioned prior it passed all 7 tests.