Reading sensor in the wrong order
Old code had it reading chnnl 1,2,3,4, and 5 where it should be
1,0,4,2,3, and 7. Resolved that issue and made w counter 1 bit as we
only needed to wait an additional 1 cycle for multiplication. Did do
math tests for Test 5 and 6, and the results are the same as calculated
with PI. So this passed all tests with PI match check. Should be able to
program/synthesize.